admin/requirements.txt
aiofile==3.8.7
aiohttp==3.8.4
aiohttp-retry==2.8.3
certifi==2023.5.7
fastapi==0.97.0
Jinja2==3.1.2
pydantic==1.10.9
python-dotenv==1.0.0
starlette==0.27.0
sqlmodel==0.0.8
urllib3==1.26.16
uvicorn==0.22.0
sqlalchemy==1.4.41
psycopg2-binary==2.9.6
requests~=2.31.0
apscheduler==3.10.4
pytz~=2024.1
beautifulsoup4~=4.12.3
sentry-sdk[fastapi]
